requirements.txt
flask
python-dotenv
supabase
gotrue
pandas
sqlalchemy
openai
pymupdf
pymysql
cryptography
gunicorn
plotly